convert builtin commands to TypeScript
