+++ src/commands/role_assign.ts
@@ -1,19 +1,20 @@
 import MessageCommand from '../lib/commands/messageCommand'
 import DatabaseManager from '../lib/databaseManager'
+import ConfigManager from '../lib/configManager'
 
 const database = new DatabaseManager()
 const { Role } = database.models
 
-export default class RoleDeassignCommand extends MessageCommand {
-  constructor(args, config) {
+export default class RoleAssignCommand extends MessageCommand {
+  constructor(args: any[], config: ConfigManager) {
     super(args, config)
 
     this.requireCommandPrefix = true
-    this.commandName = 'deassign-role'
+    this.commandName = 'assign-role'
   }
 
   static commandName() {
-    return 'deassign-role'
+    return 'assign-role'
   }
 
   command() {
@@ -31,15 +32,22 @@ export default class RoleDeassignCommand extends MessageCommand {
           throw new Error()
         }
 
-        if (this.message.member.roles.find('name', role.name)) {
+        if (role.whitelisted === false) {
+          this.message.channel.send(
+            "You can't assign that role to yourself, sorry!"
+          )
+          throw new Error()
+        }
+
+        if (!this.message.member.roles.find('name', role.name)) {
           let guildRole = this.message.guild.roles.find('name', role.name)
 
-          this.message.member.removeRole(guildRole)
+          this.message.member.addRole(guildRole)
           this.message.channel.send(
-            `Deassigned the role \`${role.name}\` from yourself!`
+            `You now have the role \`${role.name}\` assigned to yourself!`
           )
         } else {
-          this.message.channel.send('This role is not assigned to you!')
+          this.message.channel.send('This role is already assigned to you!')
         }
       })
       .catch(Error, () => {})
