+++ src/commands/role_delete.ts
@@ -5,12 +5,6 @@ const database = new DatabaseManager()
 const { Role } = database.models
 
 export default class RoleDeleteEventCommand extends RoleDeleteCommand {
-  constructor(args, config) {
-    super(args, config)
-
-    this.commandName = 'role-delete-event'
-  }
-
   static commandName() {
     return 'role-delete-event'
   }
